+++ controlplane/src/core/repositories/analytics/util.ts
@@ -368,3 +368,14 @@ export const getGranularity = (range: number) => {
 
   return '5';
 };
+
+// parse a Date to ISO9075 format in UTC, as used by Clickhouse
+export const toISO9075 = (date: Date): string => {
+  const year = date.getUTCFullYear();
+  const month = String(date.getUTCMonth() + 1).padStart(2, '0');
+  const day = String(date.getUTCDate()).padStart(2, '0');
+  const hours = String(date.getUTCHours()).padStart(2, '0');
+  const minutes = String(date.getUTCMinutes()).padStart(2, '0');
+  const seconds = String(date.getUTCSeconds()).padStart(2, '0');
+  return `${year}-${month}-${day} ${hours}:${minutes}:${seconds}`;
+};
